move remote mount to smurf and add plugin
