+++ src/simdata/__init__.py
@@ -1,2 +1,7 @@
 from .data import Data
-from .remotedata import RemoteData
+
+try:
+    from .smurf_plugin import RemoteData
+    from .smurf_plugin import SmurfData as SData
+except ImportError:
+    pass
